Move getSVGPoint function
